add a simple benchmark to compare various versions of ujson
