+++ benchmark/download_data.sh
@@ -1,0 +1,1 @@
+curl -O https://data.gharchive.org/2015-01-01-15.json.gz && gunzip 2015-01-01-15.json.gz
